[update to 4.2] read update notes or readme file
I've been without motivation for this project for quite some time now. I'm considering dropping everything and either starting anew or otherwise rebuilding it all from the ground up. The previous branch, "main" will still be hosted at the same address with the commit previous to the one this update is pushed to for the time being. HoYo have their own calculator of sorts already. Other websites will have similar tools to this.

+++ scripts/cookieLocalStorage.js
@@ -49,7 +49,42 @@ var setLSItem = (LSName, value) => {
 		console.error("Could not set "+LSName+" as the value was not a string.");
 		return;
 	}
-	localStorage.setItem(LSName,value);
+	try {
+		localStorage.setItem(LSName,value);
+	}
+	catch {
+		console.error("localStorage is full. Nothing was set. This will very likely cause problems.");
+	}
+}
+var updateLSItem = (LSName,key,newValue,mustExist=false) => {
+	// You can use undefined to delete the key
+	var item = parseLSItem(LSName,null);
+
+	if(item !== null) {
+		if(newValue === undefined) {
+			delete item[key];
+		} else {
+			item[key] = newValue;
+		}
+		if(JSON.stringify(item) === "{}") {
+			localStorage.removeItem(LSName);
+		} else {
+			setLSItem(LSName,JSON.stringify(item));
+		}
+		return;
+	} else if(!mustExist) {
+		item = new Object();
+		item[key] = newValue;
+		if(JSON.stringify(item) === "{}") {
+			localStorage.removeItem(LSName);
+		} else {
+			setLSItem(LSName,JSON.stringify(item));
+		}
+		return;
+	} else {
+		console.error("The localStorage item \""+LSName+"\" could not be updated as the \"mustExist\" flag set is set to true and the item does not exist. Aborting.");
+		return;
+	}
 }
 var getLSItem = (LSName) => {
 	return localStorage.getItem(LSName);
@@ -57,6 +92,60 @@ var getLSItem = (LSName) => {
 var clearLSItem = (LSName) => {
 	localStorage.removeItem(LSName);
 }
-var clearLS = () => {
-	localStorage.clear();
+var parseLSItem = (LSNameToGet,fallback) => {
+	// JSON.parse the localStorage item. Fallback is undefined by default.
+	if(getLSItem(LSNameToGet) != undefined) {
+		try {
+			return JSON.parse(getLSItem(LSNameToGet));
+		}
+		catch {
+			console.error("The localStorage item for \""+LSNameToGet+"\" could not be parsed. Some data may be lost. Returning fallback.");
+		}
+	}
+	return fallback;
+}
+var clearLS = (sure=false) => {
+	if(sure !== "true") {
+		console.warn("Type clearLS(\"true\") to delete EVERYTHING in localStorage.");
+		return;
+	} else {
+		localStorage.clear();
+		return;
+	}
+}
+
+var exportLS = () => {
+	console.info("Below is the localStorage as a string.");
+	console.log(JSON.stringify(localStorage));
+}
+var importLS = (sure=false,importData) => {
+	var data;
+	// deny import of null, bools and empty arrays/objects
+	if(typeof(importData) == "string") {
+		try {
+			data = JSON.parse(importData);
+		}
+		catch {
+			console.error("Could not parse import data. Aborting.");
+			return;
+		}
+		if(importData === null || importData === true || importData === false || JSON.stringify(importData) === "[]" || JSON.stringify(importData) === "{}") return;
+	} else {
+		data = importData;
+	}
+	if(Array.isArray(data)) {
+		console.error("Arrays are not allowed to be imported. Aborting.");
+		return;
+	}
+	if(sure !== "true") {
+		console.warn("Type importLS(\"true\",<data>) to import data. This WILL clear existing localStorage data.");
+		return;
+	} else {
+		clearLS("true");
+		// thanks stackoverflow q 13335967 ans 34816783
+		Object.keys(data).forEach(function (k) {
+			setLSItem(k, JSON.stringify(data[k]));
+		});
+		console.log("Your localStorage has been cleared and set to the imported data.");
+	}
 }
